+++ src/id_translation/mapping/__init__.py
@@ -2,6 +2,7 @@
 
 For and introduction to mapping, see :ref:`mapping-primer`.
 """
+import logging as _logging
 
 from ._cardinality import Cardinality
 from ._directional_mapping import DirectionalMapping
@@ -14,3 +15,7 @@ __all__ = [
     "DirectionalMapping",
     "Mapper",
 ]
+
+VERBOSE_LOGGER = _logging.getLogger(__package__).getChild("verbose")
+"""Verbose logger. Only logs messages on the ``DEBUG`` level. Disabled by default."""
+VERBOSE_LOGGER.disabled = True